% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_jack_ks.R
\name{hist_jack_ks}
\alias{hist_jack_ks}
\title{Histogram with Jackstrap Efficiency Indicators:  This function builds graphics with distributions of efficiency indicators
without outliers and complete sample. The outliers are defined by K-S Test.}
\usage{
hist_jack_ks(efficiency, model_hist_ks)
}
\arguments{
\item{efficiency}{is the jackstrap object created by jackstrap function.}

\item{model_hist_ks}{is the desired graphic model. There are four kinds: 1- Density Histogram of efficiency indicator with
complete sample and without outliers by K-S test; 2 - Histogram of efficiency with complete sample and without outliers by K-S test;
3 - Histogram of efficiency without ouliers by K-S test; 4 - Histogram of efficiency with complete sample.}
}
\value{
Return the plot with efficiency indicators with complete sample and/or without outliers by combination leverage level and K-S test;
}
\description{
Histogram with Jackstrap Efficiency Indicators:  This function builds graphics with distributions of efficiency indicators
without outliers and complete sample. The outliers are defined by K-S Test.
}
\examples{
hist_jack_ks(efficiency_ks, 1)
hist_jack_ks(efficiency_ks, 2)
hist_jack_ks(efficiency_ks, 3)
hist_jack_ks(efficiency_ks, 4)
}

