% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstrap.R
\name{jackstrap}
\alias{jackstrap}
\title{Jackstrap Method: The tool identifies outliers in Nonparametric Frontier.
This function applies the developed tecnique by Sousa and Stosic (2005)
Technical Efficiency of the Brazilian Municipalites: Correcting Nonparametric
Frontier Meansurements for Outliers}
\usage{
jackstrap(data, ycolumn, xcolumn, bootstrap = 1000,
  perc_sample_buble = 0.2, dea_method = "vrs",
  orientation_dea = "in", n_seed = NULL, repos = FALSE,
  num_cores = 1)
}
\arguments{
\item{data}{is the dataset with input and output used to mensure efficiency; Dataset needs to have this form: 1th column: name of DMU (string); 2th column: code of DMU (integer); n columns of output variables; n columns of input variables.}

\item{ycolumn}{is the quantity of y columns of dataset.}

\item{xcolumn}{is the quantity of x columns of dataset.}

\item{bootstrap}{is the quantity of applied resampling.}

\item{perc_sample_buble}{is the percentage of sample in each bubble.}

\item{dea_method}{is the DEA method: "crs" is DEA with constant returns to scale (CCR); "vrs" is DEA with varible returns to scale; and "fdh" is Free Disposal Hull (FDH) with variable returns to scale.}

\item{orientation_dea}{is the direction of the DEA: "in" for focus on inputs; and "out" for focus on outputs.}

\item{n_seed}{is the code as seed used to get new random samples.}

\item{repos}{identify if the resampling method is with replacement TRUE or not FALSE.}

\item{num_cores}{is the number of cores available to process.}
}
\value{
Return the jackstrap object with follow informations: "parameters" contain the parameters used on function; "bootstrap" is quantity of applied resampling;
"mean_leverage" is leverage average for each DMU; "mean_geral_leverage" is general average of leverage and step function threshold;
"sum_leverage" is accrued leverage on all resampling for each DMU; "count_dmu" is amount of each DMU selected by bootstrap.
"efficiency_step_func" are efficiency indicators obtained by heaviside step function criteria; "result_kstest_method" are p-values of K-S test obtained by removing sequencially one by one the high leverage DMU;
"efficiency_ks_method" are efficiency indicators obtained by K-S test criteria.
}
\description{
Jackstrap Method: Tool identifies outliers in Nonparametric Frontier.
This function applies the developed tecnique by Sousa and Stosic (2005)
Technical Efficiency of the Brazilian Municipalites: Correcting Nonparametric
Frontier Meansurements for Outliers
}
\examples{
municipalities <- as.data.frame(jackstrap::municipalities)

efficiency <- jackstrap (data=municipalities, ycolumn=2, xcolumn=1, bootstrap=500,
                         perc_sample_bubble=0.10, dea_method="vrs",
                         orientation_dea="in", n_seed = 2000, repos=FALSE, num_cores=4)
}
