% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstrap_ks.R
\name{jackstrap_ks}
\alias{jackstrap_ks}
\title{Jackstrap KS Method: The tool identifies outliers in Nonparametric Frontier.
This function applies the developed tecnique by Sousa and Stosic (2005)
Technical Efficiency of the Brazilian Municipalites: Correcting Nonparametric
Frontier Meansurements for Outliers and to use the K-S test with criteria to define outliers}
\usage{
jackstrap_ks(data, jackstrap_obj, num_cores, perc)
}
\arguments{
\item{data}{is the dataset with input and output used to mensure efficiency; Dataset need to have this form: 1th column: name of DMU (string); 2th column: code of DMU (integer); n columns of output variables; n columns of input variables.}

\item{jackstrap_obj}{is the object created by the function jackstrap.}

\item{num_cores}{is number of cores used to process.}

\item{perc}{is the percentage of sample used to K-S test.}
}
\value{
Return the jackstrap object increased with follow informations: "result_kstest_method" are p-values of K-S test obtained by removing sequencially one by one the high leverage DMU;
"efficiency_ks_method" are efficiency indicators obtained by K-S test criteria.
}
\description{
Jackstrap KS Method: Tool identifies outliers in Nonparametric Frontier.
This function applies the developed tecnique by Sousa and Stosic (2005)
Technical Efficiency of the Brazilian Municipalites: Correcting Nonparametric
Frontier Meansurements for Outliers and to use the K-S test with criteria to define outliers
}
\examples{
efficiency_ks <- jackstrap_ks (data=municipalities, jackstrap_obj=efficiency, num_cores = 4, perc=0.80)
}

