% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jackstrap_ks.R
\name{plot_jackstrap_ks}
\alias{plot_jackstrap_ks}
\title{Plot Jackstrap KS:  This function plots p-value of Kolmogorov-Smirnov Test in order decresing of leverage.}
\usage{
plot_jackstrap_ks(data_plot, model_plot)
}
\arguments{
\item{data_plot}{is the jackstrap object created by jackstrap function.}

\item{model_plot}{is the desired model. There are two models: 1 - The graphic shows the amount of removed DMU on x axis and p-value of K-S test on y axis;
2 - The graphic shows DMU code on x axis and p-value of K-S test on y axis.}
}
\value{
Return the plot with p-value of K-S test and removed DMU or DMU code.
}
\description{
Plot Jackstrap KS:  This function plots p-value of Kolmogorov-Smirnov Test in order decresing of leverage.
}
\examples{
plot_jackstrap_ks(efficiency_ks, 1)
}
