% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_jackstrap.R
\name{summary_jackstrap}
\alias{summary_jackstrap}
\title{Summary Jackstrap:  This function shows the main outcomes with outlier technique developed by
Sousa and Stosic(2005).}
\usage{
summary_jackstrap(object_jackstrap)
}
\arguments{
\item{object_jackstrap}{is the jackstrap object created by jackstrap function.}
\item{data}{is the dataset of research.}
}
\value{
Return the data frame with follow informations: "parameters" contain the parameters used on function;
"outliers_by_step_func" are the outliers by heaviside step function criteria; "outliers_by_ks" are the outliers by K-S test;
"dmu_efficiency_by_step_func" are DMUs evaluated as efficients by heaviside step function criteria; "dmu_inefficiency_by_step_func" are the DMUs evaluated as maximum inefficient by heaviside step function criteria;
"dmu_efficiency_ks" are DMUs evaluated as efficients by K-S test criteria; "dmu_inefficiency_by_ks" are the DMUs evaluated as maximum inefficient by K-S test criteria.
}
\description{
Summary Jackstrap:  This function shows the main outcomes with outlier technique developed by
Sousa and Stosic(2005).
}
\examples{
summary_efic <- summary_jackstrap(efficiency_ks)
}

