% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_jack_step.R
\name{hist_jack_step}
\alias{hist_jack_step}
\title{Histogram with Jackstrap Efficiency Indicators:  This function builds a graphic with indicator distributions
without outliers and complete sample. The outliers are defined by heaviside step function.}
\usage{
hist_jack_step(efficiency, model_hist_step)
}
\arguments{
\item{efficiency}{is the jackstrap object created by jackstrap function.}

\item{model_hist_step}{is the desired graphic model. There are four kinds: 1- Density Histogram of efficiency indicators with
complete sample and without outliers by heaviside step function; 2 - Histogram of efficiency with complete sample and without outliers
by heaviside step function; 3 - Histogram of efficiency without ouliers by heaviside step function; 4 - Histogram of efficiency with complete sample.}
}
\value{
Return the plot with efficiency indicators with complete sample and/or without outliers by heaviside step function;
}
\description{
Histogram with Jackstrap Efficiency Indicators:  This function builds a graphic with indicator distributions
without outliers and complete sample. The outliers are defined by heaviside step function method.
}
\examples{
hist_jack_step(efficiency, 1)
hist_jack_step(efficiency, 2)
hist_jack_step(efficiency, 3)
hist_jack_step(efficiency, 4)
}
